Game API v1.0.38: Stream abstraction
This changes the Game API to use a common abstraction for audio and video
streams, in preparation for framebuffer streams.

The primary benefit is that video metadata is passed to the frontend
in-band along with the data. Previously, metadata was only specified on
stream open, requiring a new stream if metadata changes. Now, the common
stream struct puts the data alongside its metadata.

RetroPlayer has been extended so that it can scale to hardware and software
streams in the future.